requirements.txt
openai 
flask==2.3.3
flask-cors==4.0.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
numpy>=1.21.0 